adjusted orgnization
