+++ tripapi/serializers/__init__.py
@@ -1,4 +1,5 @@
 from .user_serializer import UserSerializer
 from .trip_serializer import TripSerializer
 from .invite_serializer import InviteSerializer
-from .event_serializer import EventSerializer, LikeSerializer
+from .event_serializer import EventSerializer
+from .like_serializer import LikeSerializer
